mnt/data/proj_mod/requirements.txt
telethon>=1.36.0
python-telegram-bot>=21.0
